+++ catalystwan/endpoints/configuration/policy/list/local_app.py
@@ -5,17 +5,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import LocalAppList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.local_app import LocalAppList, LocalAppListEditPayload, LocalAppListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class LocalAppListEditPayload(LocalAppList, PolicyListId):
-    pass
-
-
-class LocalAppListInfo(LocalAppList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyLocalAppList(APIEndpoints, PolicyListEndpoints):
