+++ catalystwan/models/policy/__init__.py
@@ -18,6 +18,7 @@ from catalystwan.models.policy.list.fqdn import FQDNList
 from catalystwan.models.policy.list.geo_location import GeoLocationList
 from catalystwan.models.policy.list.ips_signature import IPSSignatureList
 from catalystwan.models.policy.list.ipv6_prefix import IPv6PrefixList
+from catalystwan.models.policy.list.local_app import LocalAppList
 
 from .centralized import CentralizedPolicy, TrafficDataDirection
 from .definition.access_control_list import AclPolicy
@@ -35,7 +36,6 @@ from .definition.traffic_data import TrafficDataPolicy
 from .definition.vpn_membership import VPNMembershipPolicy
 from .definition.zone_based_firewall import ZoneBasedFWPolicy
 from .lists import (
-    LocalAppList,
     LocalDomainList,
     MirrorList,
     PolicerList,
